Added a mapping function (exponent)
added -e / --exponent.
As the perceived brightness change from 10% to 20% is not the same as 90% to 100%,
there is now a float parameter--exponent / -e to set the mapping of the percentage form linear (as it was) to polynominal.
(%=x^k*max*100^-k)
k is the value that you can change now.