Use non-suid permissions when logind is used
When systemd-logind and its D-Bus API is used to actual change the
brightness of devices the binary does not have to be suid, which is
desired for security reasons.

This is a cherry-pick of: 7504b58765cafae7f7628fe02204883a200a4b10 that
was partially reverted by: 052e56ccd662fba25abd7780be5feb66a8e0710b.

Signed-off-by: Antoine Damhet <antoine.damhet@lse.epita.fr>